Consolidate type definitions into single source of truth
- Merge Session, Decision, Task, and CurrentFocus types from
  src/context-bridge/types.ts into src/types.ts
- Decision: add both date and timestamp fields, expand madeBy to include
  sonnet/haiku, add both reversible and confidence
- Task: include dependencies/artifacts arrays, add inline execution
  metadata (model, duration, cost)
- Session: add modelsInvolved tracking
- CurrentFocus: make lastUpdated required
- Add TaskResult and ContextBridgeConfig from context-bridge
- Update all imports to use consolidated types
- Delete redundant src/context-bridge/types.ts
